requirements.txt
numpy~=1.20.1
ipywidgets~=7.6.3
pillow~=8.2.0
setuptools~=52.0.0
matplotlib~=3.3.4